GLT-4432_changes_requested_in_PR

+++ miso-web/src/main/webapp/scripts/bulk_library.js
@@ -960,7 +960,7 @@ BulkTarget.library = (function ($) {
           title: "Spike-In Volume",
           type: "decimal",
           data: "spikeInVolume",
-          precision: 16,
+          precision: 14,
           scale: 10,
         }
       );
